+++ frontend/src/integrations/github/GithubIntegration.ts
@@ -1,36 +1,14 @@
 import reposJsonFixture from './repos.json'
+import {
+  SortOrder,
+  SortColumns,
+  type GithubRepos,
+  type GithubLangBreakdown,
+  type LanguageData,
+  LOWERCASE_SORT_COLUMNS
+} from './GithubTypes'
 
-export interface GithubRepoLangBreakdown { [key: String]: number }
-
-export interface GithubRepo {
-  created_at: Date,
-  description: String,
-  html_url: String,
-  language: String,
-  languageBreakdown: GithubRepoLangBreakdown,
-  languageData: LanguageData
-  name: String,
-  pushed_at: Date,
-  updated_at: Date,
-}
-
-export type GithubRepos = GithubRepo[]
-
-export enum SortOrder {
-  ASC = "asc",
-  DESC = "desc"
-}
-export enum SortColumns {
-  NAME = "name",
-  LANGUAGE = "language",
-  DESCRIPTION = "description",
-  PUSHED_AT = "pushed_at",
-  UPDATED_AT = "updated_at",
-  CREATED_AT = "created_at"
-}
-const LOWERCASE_SORT_COLUMNS = [SortColumns.NAME, SortColumns.DESCRIPTION]
-
-export class GithubIntegration {
+export default class GithubIntegration {
   sortOrder: SortOrder = SortOrder.DESC
   sortColumn: SortColumns = SortColumns.PUSHED_AT
   repos: Writeable<GithubRepos>
@@ -49,7 +27,7 @@ export class GithubIntegration {
     this.updateRepos = (val: boolean) => updateRepos(val)
   }
 
-  languageBreakdownToData(breakdown: GithubRepoLangBreakdown|null): LanguageData|void {
+  languageBreakdownToData(breakdown: GithubRepoLangBreakdown|null|undefined): LanguageData|void {
     if (breakdown)
     {
       var data: LanguageData
